Fixes for working model & mappings

+++ BooksAndMore.Catalogue/Infrastructure/BooksAndMore.Catalogue.Infrastructure.Data/Mapping/ValueGenerators/AverageRatingValueGenerator.cs
@@ -5,11 +5,11 @@ using System.Linq;
 
 namespace BooksAndMore.Catalogue.Infrastructure.Data.Mapping.ValueGenerators
 {
-    public class AverageRatingValueGenerator : ValueGenerator<double>
+    public class AverageRatingValueGenerator : ValueGenerator<decimal>
     {
         public override bool GeneratesTemporaryValues => false;
 
-        public override double Next(EntityEntry entry)
+        public override decimal Next(EntityEntry entry)
         {
             if (entry == null)
             {
@@ -18,7 +18,7 @@ namespace BooksAndMore.Catalogue.Infrastructure.Data.Mapping.ValueGenerators
 
             var book = entry.Entity as Book;
             return book != null ?
-                book.Reviews.Select(review => review.Rating).DefaultIfEmpty(0).Average() :
+                (decimal)book.Reviews.Select(review => review.Rating).DefaultIfEmpty(0).Average() :
                 0;
         }        
     }
